+++ api_tests/btsieve/bitcoin_only/test.ts
@@ -103,86 +103,46 @@ setTimeout(async function() {
                     );
                 });
 
-                it("btsieve should respond with no content when deleting an existing bitcoin transaction query", async function() {
-                    const res = await request(
-                        btsieve.absoluteLocation(location)
-                    ).del("");
+                it('btsieve should respond with 200 OK when "getting-or-creating" a query that already exists', async function() {
+                    const res = await request(btsieve.url())
+                        .put(location)
+                        .send({
+                            to_address: toAddress,
+                        });
+
+                    expect(res).to.have.status(200);
+                });
+
+                it('btsieve should respond with 409 CONFLICT when "getting-or-creating" an existing query ID with a different query body', async function() {
+                    const differentToAddress =
+                        "mzkdMKoki1hoP3ogT2oGSJ4pBTC9UGDLfM";
+
+                    const res = await request(btsieve.url())
+                        .put(location)
+                        .send({
+                            to_address: differentToAddress,
+                        });
+
+                    expect(res).to.have.status(409);
+                });
+
+                it('btsieve should respond with 204 NO_CONTENT when "getting-or-creating" a new query', async function() {
+                    const newQueryId =
+                        "4BvFBixM4HmhV8AJe5RC8v8csxxhDBscMxwpiK5e";
+                    const newLocation =
+                        "/queries/bitcoin/regtest/transactions/" + newQueryId;
+                    const newQuery = {
+                        to_address: "mndvsV4weBdPFTarHQbg71rCXRy8z79SH5",
+                    };
+
+                    const res = await request(btsieve.url())
+                        .put(newLocation)
+                        .send(newQuery);
 
                     expect(res).to.have.status(204);
                 });
-            });
 
-            describe("Blocks", () => {
-                it("btsieve should respond not found when getting a non-existent bitcoin block query", async function() {
-                    const res = await request(btsieve.url()).get(
-                        "/queries/bitcoin/regtest/blocks/1"
-                    );
-
-                    expect(res).to.have.status(404);
-                });
-
-                const minHeight = 200;
-                let location: string;
-                it("btsieve should respond not found when creating a bitcoin block query for an invalid network", async function() {
-                    const res = await request(btsieve.url())
-                        .post("/queries/bitcoin/banananet/blocks")
-                        .send({
-                            min_height: minHeight,
-                        });
-
-                    expect(res).to.have.status(404);
-                });
-
-                it("btsieve should respond with location when creating a valid bitcoin block query", async function() {
-                    const res = await request(btsieve.url())
-                        .post("/queries/bitcoin/regtest/blocks")
-                        .send({
-                            min_height: minHeight,
-                        });
-
-                    location = res.header.location;
-
-                    expect(res).to.have.status(201);
-                    expect(location).to.not.be.empty;
-                });
-
-                it("btsieve should respond with no match when querying an existing bitcoin block query", async function() {
-                    const res = await request(
-                        btsieve.absoluteLocation(location)
-                    ).get("");
-
-                    expect(res).to.have.status(200);
-                    expect(res.body.query.min_height).to.equal(minHeight);
-                    expect(res.body.matches).to.be.empty;
-                });
-
-                it("btsieve should respond with no block match (yet) when requesting on the min_height 600 bitcoin block query", async function() {
-                    this.slow(500);
-                    await bitcoin.generate(50);
-
-                    const res = await request(
-                        btsieve.absoluteLocation(location)
-                    ).get("");
-
-                    expect(res).to.have.status(200);
-                    expect(res.body.query.min_height).to.equal(minHeight);
-                    expect(res.body.matches).to.be.empty;
-                });
-
-                it("btsieve should respond with block match when requesting on the min_height 600 bitcoin block query", async function() {
-                    this.slow(2000);
-                    this.timeout(3000);
-
-                    await bitcoin.generate(50);
-                    const body = await btsieve.pollUntilMatches<IdMatch>(
-                        btsieve.absoluteLocation(location)
-                    );
-
-                    expect(body.query.min_height).to.equal(minHeight);
-                    expect(body.matches).to.have.length.greaterThan(1);
-                });
-
-                it("btsieve should respond with no content when deleting an existing bitcoin block query", async function() {
+                it("btsieve should respond with no content when deleting an existing bitcoin transaction query", async function() {
                     const res = await request(
                         btsieve.absoluteLocation(location)
                     ).del("");