+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- All the code since the dawn of comit-rs
+- All the code since the dawn of comit-rs.
+
+### Changed
+- Move config files to standard location based on platform (OSX, Windows, Linux).
 
 [Unreleased]: https://github.com/comit-network/comit-rs/compare/1625533e04119e8496b14d5e18786f150b4fce4d...HEAD
-